:zap: (feature/edit-profile) 스택뷰의 arrangedSubviews 를 모두 지우는 동작을 함수화 하여 코드 리팩토링 수행
